Optimized ModalStyle.scss with SCSS nesting, added TODO comments for future refinements
