+++ README.md
@@ -1,1 +1,21 @@
+# Pomo - Pomodoro Timer App
 
+**Version:** 0.1.0
+
+## Overview
+Pomo is a user-friendly Pomodoro timer designed to enhance productivity through time management. 
+The app helps users stay focused and manage their work and break sessions effectively.
+
+## Features
+- **Focus Timer:** A dedicated timer for focused work sessions, defaulting to 25 minutes.
+- **Short Break Timer:** A quick break timer, set to 5 minutes, to recharge between work sessions.
+- **Long Break Timer:** A longer break timer, set to 15 minutes, ideal after multiple work sessions.
+- **Simple Controls:** Intuitive play, pause, and reset buttons for easy session management.
+- **Responsive Design:** Optimized for both desktop and mobile, providing a seamless experience on any device.
+
+## Usage
+Pomo is designed to support users in their work and break cycles, improving productivity and helping to prevent burnout.
+
+---
+
+Feel free to clone, fork, and contribute to the project to make it even better!
