+++ index.html
@@ -3,7 +3,7 @@
 
 <head>
   <meta charset="UTF-8" />
-  <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+  <link rel="icon" type="image/svg+xml" href="./src/assets/icons/icoBackUp/clock.svg" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
   <link rel="preconnect" href="https://fonts.googleapis.com">
@@ -13,7 +13,7 @@
     rel="stylesheet">
 
   <link rel="stylesheet" href="./src/assets/icons/icomoon.css">
-  <title>Vite + React + TS</title>
+  <title>Pomo Timer</title>
 </head>
 
 <body>
