+++ src/assets/color-variables.scss
@@ -2,10 +2,10 @@
 
 // light theme //
 
-$phocus-light-background: #fff2f2;
-$phocus-light-switch-btn: #ff7c7c;
-$phocus-light-secondary-btn: #ffd9d9;
-$phocus-light-text-color: #471515;
+$focus-light-background: #fff2f2;
+$focus-light-switch-btn: #ff7c7c;
+$focus-light-secondary-btn: #ffd9d9;
+$focus-light-text-color: #471515;
 
 $long-break-light-background: #f2f9ff;
 $long-break-light-switch-btn: #8bcaff;
@@ -19,10 +19,10 @@ $short-break-light-text-color: #14401d;
 
 // dark theme //
 
-$phocus-dark-background: #0d0404;
-$phocus-dark-switch-btn: #b93737;
-$phocus-dark-secondary-btn: #310e0e;
-$phocus-dark-text-color: #fff2f2;
+$focus-dark-background: #0d0404;
+$focus-dark-switch-btn: #b93737;
+$focus-dark-secondary-btn: #310e0e;
+$focus-dark-text-color: #fff2f2;
 
 $short-break-dark-background: #040d06;
 $short-break-dark-switch-btn: #328c46;
