+++ src/components/InfoBlock/InfoBlock.scss
@@ -1,4 +1,4 @@
-@import '../../assets/variables.scss';
+@import '../../assets/color-variables.scss';
 
 .info-block {
   display: flex;
@@ -17,7 +17,7 @@
   .logo-text {
     font-size: 65px;
     font-weight: 800;
-    color: $phocus-light-text-color;
+    color: $focus-light-text-color;
   }
 
   .logo-icon {
@@ -30,11 +30,17 @@
   }
 }
 
+//TODO: BEM rename like -
+// .logo-block {
+//   &__text { ... }
+//   &__icon { ... }
+// }
+
 .intro-block {
   .logo-title {
     font-size: 40px;
     font-weight: 900;
-    color: $phocus-light-text-color;
+    color: $focus-light-text-color;
     margin-top: 70px;
     margin-bottom: 0;
   }
