+++ src/main.tsx
@@ -2,9 +2,9 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 
 import App from './App.tsx'
-import { FocusTimer } from './timers/FocusTimer.tsx'
-import { LongBreak } from './timers/LongBreak.tsx'
-import { ShortBreak } from './timers/ShortBreak.tsx'
+import { FocusTimer } from './timers/FocusTimer/FocusTimer.tsx'
+import { LongBreak } from './timers/LongBreak/LongBreak.tsx'
+import { ShortBreak } from './timers/ShortBreak/ShortBreak.tsx'
 
 import { createBrowserRouter, RouterProvider} from 'react-router-dom'
 
@@ -15,15 +15,15 @@ const router = createBrowserRouter([
   },
   {
     path:"/FocusTimer",
-    element:<FocusTimer />,
+    element:<FocusTimer defaultFocusTime={0} />,
   },
   {
     path:"/LongBreak",
-    element:<LongBreak />,
+    element:<LongBreak defaultLongBreak={0}/>,
   },
   {
     path:"/ShortBreak",
-    element:<ShortBreak />,
+    element:<ShortBreak defaultShortBreak={0}/>,
   },
 ])
