--- src/timers/FocusTimer.scss
@@ -1,48 +1,0 @@
-@import '../assets/variables.scss';
-
-.focus-background {
-  background-color: $phocus-light-background;
-}
-
-.focus-main-container {
-  background-color: $phocus-light-background;
-}
-
-.focus-mode__btn {
-  background-color: $phocus-light-secondary-btn;
-  border: 2px solid $phocus-light-text-color;
-  width: 120px;
-}
-
-.focus-mode__ico {
-  background-image: url(../assets/icons/IcoBackUp/brain.svg);
-  width: 25px;
-  height: 25px;
-  background-size: contain;
-}
-
-.focus-mode__text {
-  color: $phocus-light-text-color;
-}
-
-.focus-timer {
-  color: $phocus-light-text-color;
-}
-
-.focus-menu__btn {
-  margin-right: 15px;
-  background-color: $phocus-light-secondary-btn;
-}
-
-.focus-forward__btn {
-  margin-left: 15px;
-  background-color: $phocus-light-secondary-btn;
-}
-
-.focus-start__btn {
-  background-color: $phocus-light-switch-btn;
-}
-
-.btn-pressed {
-  background-color: gray;
-}
